refactor(projectDetailsV2): update folder & file name from ProjectDetailsV2 to ProjectDetails
