+++ src/frontend/src/components/ProjectDetails/FeatureSelectionPopup.tsx
