+++ src/frontend/src/components/ProjectDetails/MobileFooter.tsx
