+++ src/frontend/src/components/ProjectDetails/MobileProjectInfoContent.tsx
