+++ src/frontend/src/components/ProjectDetails/ProjectInfo.tsx
