+++ src/frontend/src/components/ProjectDetails/ProjectOptions.tsx
