+++ src/frontend/src/components/ProjectDetails/SkeletonLoader.tsx
